cast to float32

+++ model.py
@@ -67,8 +67,8 @@ class Model(object):
         # self.resized_x = tf.image.resize_images(self.x, resize_dim)
         # self.resized_y = tf.image.resize_images(self.y, resize_dim)
         # self.resized_test_x = tf.image.resize_images(self.test_x, resize_dim)
-        self.resized_x = self.x
-        self.resized_y = self.y
+        self.resized_x = tf.cast(self.x, tf.float32)
+        self.resized_y = tf.cast(self.y, tf.float32)
 
         self.normalized_x = normalize(self.resized_x)
         self.normalized_y = normalize(self.resized_y)
